+++ GPT-VS-EMPTY-CUDA/main.h
@@ -4,4 +4,4 @@
 #include "stdInte.h"
 
 
-void initGpt2(double gpt[3][3], double alpha, double beta, double b1, double b2, double rotation);
+void initGpt2(float gpt[3][3], float alpha, float beta, float b1, float b2, float rotation);
